requirements.txt
Django==4.0
djangorestframework==3.13.1
requests==2.27.1
redis==4.1.3
celery==5.2.3
psycopg2==2.9.3
django-celery-results==2.2.0
django-filter==21.1
django-bootstrap4==21.2
django-tables2==2.4.1
pillow==9.0.1
django-model-utils==4.2.0
django-cors-headers==3.6.0
Faker==13.3.4